Session 2 + Session additional task;

+++ Desktop/RealEstateAgency/Core/DGridGetDeleteEntities.cs
@@ -1,5 +1,7 @@
-﻿using RealEstateAgency.Data;
+﻿using RealEstateAgency;
+using RealEstateAgency.Data;
 using System;
+using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
@@ -15,13 +17,67 @@ namespace RealEstateAgency
     class DGridEntityManager<TEntity> where TEntity : class
     {
         private readonly DataGrid _entitiesDGrid;
-        private readonly IFilter _filter;
-        public DGridEntityManager(DataGrid entitiesDGrid, IFilter filter)
+        private readonly UserErrorCheack[] _userErrorCheacks;
+        private Dictionary<string, FilterWithPattern> usingFilters;
+
+        public DGridEntityManager(DataGrid entitiesDGrid, UserErrorCheack[] userErrors)
         {
             var data = AgencyModel.Instance.Set<TEntity>().ToList();
              _entitiesDGrid = entitiesDGrid;
             _entitiesDGrid.ItemsSource = data;
-            _filter = filter;
+            _userErrorCheacks = userErrors;
+
+            usingFilters = new Dictionary<string, FilterWithPattern>();
+        }
+
+        public DGridEntityManager(DataGrid entitiesDGrid) : this(entitiesDGrid, new UserErrorCheack[0])
+        {
+        }
+
+        public IEnumerable<TEntity> AllEntities => AgencyModel.Instance.Set<TEntity>().ToList();
+        public IEnumerable<TEntity> DisplayedEntities => (IEnumerable<TEntity>)_entitiesDGrid.ItemsSource;
+
+        /// <summary>
+        /// Использование фильтрации для элементов DataGrid.
+        /// Добавляет фильтр в коллекцию фильтров и применяет его если фильтра с указаным именем еще не пременено.
+        /// Если фильтр с указаным именем существует, заменяет его.
+        /// </summary>
+        /// <param name="filterName">Название фильтра</param>
+        /// <param name="filter"></param>
+        /// <param name="pattern"></param>
+        public void UseFilter(string filterName, ICollectionFilter filter, string pattern)
+        {
+            var filterForUse = new FilterWithPattern { Pattern = pattern, Filter = filter };
+
+            if(!usingFilters.ContainsKey(filterName))
+            {
+                usingFilters.Add(filterName, new FilterWithPattern { Pattern = pattern, Filter = filter });
+            }
+            else
+            {
+                usingFilters[filterName] = filterForUse;
+            }
+            ReloadTable();
+        }
+
+        public bool IsFilterUsing(string filterName)
+        {
+            return usingFilters.ContainsKey(filterName);
+        }
+        public FilterWithPattern GetUsingFilter(string filterName)
+        {
+            return usingFilters[filterName];
+        }
+        public void RemoveFilter(string filterName)
+        {
+            usingFilters.Remove(filterName);
+            ReloadTable();
+        }
+
+        public void ClearFilters()
+        {
+            usingFilters.Clear();
+            ReloadTable();
         }
 
         /// <summary>
@@ -43,6 +99,17 @@ namespace RealEstateAgency
             {
                 try
                 {
+                    var errors = new ErrorBuilder();
+                    foreach(var userErrorCheack in _userErrorCheacks)
+                    {
+                        errors.AssertError(userErrorCheack.СheckFunc(), userErrorCheack.Message);
+                    }
+
+                    if(errors.IsAnyError())
+                    {
+                        throw new UserActionException(errors.GetMessage());
+                    }
+
                     AgencyModel.Instance.Set<TEntity>().RemoveRange(selectedItems);
                     AgencyModel.Instance.SaveChanges();
                     
@@ -64,17 +131,17 @@ namespace RealEstateAgency
         /// </summary>
         public void ReloadTable()
         {
-            AgencyModel.Instance.ChangeTracker.Entries().ToList().ForEach(item => item.Reload());
-            _entitiesDGrid.ItemsSource = AgencyModel.Instance.Set<TEntity>().ToList();
+            ReloadTableWithoutFilters();
+            foreach(var item in usingFilters)
+            {
+                _entitiesDGrid.ItemsSource = item.Value.UseFilter(_entitiesDGrid.ItemsSource);
+            }
         }
 
-        /// <summary>
-        /// Применят фильтр к таблице.
-        /// </summary>
-        public void UseFilter(string pattern)
+        private void ReloadTableWithoutFilters()
         {
-            _entitiesDGrid.ItemsSource = _filter.Filter(AgencyModel
-                .Instance.Set<TEntity>().ToList(), pattern).Cast<TEntity>();
+            AgencyModel.Instance.ChangeTracker.Entries().ToList().ForEach(item => item.Reload());
+            _entitiesDGrid.ItemsSource = AgencyModel.Instance.Set<TEntity>().ToList();
         }
     }
 }
